Added main for generating pgm image
This commit filesystem and lwIP BSP was added. Filesystem is used for
saving image to the boards SD card.

I modified loop for saving picture and added check to ensure it is
working. Configuration is still open problem but now it is working.

Code for initializing FS and saving image is main but that will be
cleaned in the following commits.

I removed function calls from gpio_helpers as they are now inlined. From
the faculty webpage I added iic_helpers.

Signed-off-by: Darko Janekovic <darko.janekovic@fer.hr>