Refactor ov7670 interface and ffmpeg command
Argh, stupid bugs and pixel arrangement. This code produces bin that can
be compressed to JPEG using:

$ ffmpeg -s 640x480 -pix_fmt uyvy422 -i vrata422.yuv slika.jpg

Picture should be in yuyv422 format but it is in uyuv422... I don't know
what is going on but I am commiting so I can go back to this commit and
test stuff. Filesystem stuff is maybe wrong...

Signed-off-by: Darko Janekovic <darko.janekovic@fer.hr>